+++ remaku/views/components/toolbar.py
@@ -102,7 +102,6 @@ class Toolbar(QWidget):
                 {"id": "export_macro", "label": self.tr("Export Macro")},
                 {"id": "open_macro_folder", "label": self.tr("Open Macro Folder")},
                 {"separator": True},
-                {"id": "macro_explorer", "label": self.tr("Macro Explorer")},
                 {"id": "settings", "label": self.tr("Settings"), "shortcut": "Ctrl+,"},
                 {"separator": True},
                 {"id": "quit", "label": self.tr("Quit")},